processor/requirements.txt
fastapi
uvicorn[standard]
pyyaml
httpx
pydantic
